Improve Backup Logic in update-app-config.sh
This commit updates the script `update-app-config.sh` to enhance the backup mechanism for the configuration file. The modification involves adding a conditional check before creating a backup of `config.ini`. Now, the script will first verify if `config.ini` exists before attempting to back it up. This change prevents unnecessary backup operations and ensures that the backup process is more reliable and context-aware.

+++ e2e/update-app-config.sh
@@ -13,8 +13,10 @@ sed -i 's/mysql:host=127.0.0.1/mysql:host=localhost/' config.ini.temp
 # Comparing config.ini.temp to config.ini
 if ! cmp -s config.ini.temp config.ini; then
     # If they are different, back up config.ini with a date stamp
-    cp config.ini config.ini.backup_$(date +%Y%m%d_%H%M%S)
-
+    # if config.ini exists say hi
+    if [ -f config.ini ]; then
+        cp config.ini config.ini.backup_$(date +%Y%m%d_%H%M%S)
+    fi
     # Move config.ini.temp to config.ini
     mv config.ini.temp config.ini
 else
